Added spell 1, added "projectiles", need to work on ennemy death, it doesn't work (var "exist" vary without explication)

+++ src/basic.c
@@ -9,7 +9,7 @@ int aleatoire(int min, int max)
 
 int distancepoint(int x1, int y1, int x2, int y2)
 {
-    double distance;
+    int distance;
     distance=sqrt((x1-x2)*(x1-x2)+(y1-y2)*(y1-y2));
-    return (int)distance;
+    return distance;
 }
